requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
databases==0.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-mail==1.4.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pdoc3==0.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pytest==7.4.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
